chore: move lib to builders
