+++ packages/styled/src/builders/lib.ts
@@ -1,5 +1,5 @@
 import { Key } from '@charcoal-ui/theme'
-import { unreachable } from './util'
+import { unreachable } from '../util'
 
 /**
  * 配列で指定したプロパティを動的に生やす
